+++ src/main/java/com/payment/payfareserver/entity/Trip.java
@@ -25,7 +25,7 @@ public class Trip {
     private LocalDateTime date;
 
     @Column(name = "price", nullable = false, precision = 15, scale = 2)
-    private BigDecimal price;
+    private Double price;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "admin_id", nullable = false)
@@ -77,11 +77,11 @@ public class Trip {
         this.date = date;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
